+++ packages/dataviews/src/constants.js
@@ -16,19 +16,21 @@ import ViewTable from './view-table';
 import ViewGrid from './view-grid';
 import ViewList from './view-list';
 
-// Field types.
-export const ENUMERATION_TYPE = 'enumeration';
-
 // Filter operators.
 export const OPERATOR_IS = 'is';
 export const OPERATOR_IS_NOT = 'isNot';
 export const OPERATOR_IS_ANY = 'isAny';
 export const OPERATOR_IS_NONE = 'isNone';
+export const OPERATOR_IS_ALL = 'isAll';
+export const OPERATOR_IS_NOT_ALL = 'isNotAll';
+
 export const ALL_OPERATORS = [
 	OPERATOR_IS,
 	OPERATOR_IS_NOT,
 	OPERATOR_IS_ANY,
 	OPERATOR_IS_NONE,
+	OPERATOR_IS_ALL,
+	OPERATOR_IS_NOT_ALL,
 ];
 export const OPERATORS = {
 	[ OPERATOR_IS ]: {
@@ -46,6 +48,14 @@ export const OPERATORS = {
 	[ OPERATOR_IS_NONE ]: {
 		key: 'is-none-filter',
 		label: __( 'Is none' ),
+	},
+	[ OPERATOR_IS_ALL ]: {
+		key: 'is-all-filter',
+		label: __( 'Is all' ),
+	},
+	[ OPERATOR_IS_NOT_ALL ]: {
+		key: 'is-not-all-filter',
+		label: __( 'Is not all' ),
 	},
 };