+++ packages/react-native-editor/CHANGELOG.md
@@ -10,6 +10,11 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+-   [*] Add empty fallback option for the BottomSheetSelectControl component [#60333]
+-   [*] Fix Quote Block styles [#60476]
+
+## 1.116.0
+-   [**] Highlight color formatting style improvements [#57650]
 
 ## 1.115.0
 -   [*] Improve consistency of the block outline indicating the currently selected block [#59415]